started licenses and give runs custom names

+++ electron/app/index.js
@@ -65,6 +65,7 @@ let vm = new Vue({
             "singleFC-ptm": {"text": "Single fold change PTM", "er": true},
             "multiFC-ptm": {"text": "Multi fold change PTM", "er": true},
         },
+        extra_required_fields: ["mods", "fasta_file", "enzyme"],
         species_map: {"homo sapiens": "human", "mus musculus": "mouse", "rattus norvegicus": "rat"},
         allowed_runs: ["string", "genemania", "both"],
         allowed_visualize: ["biological process","cellular component","molecular function","pathways","all"],
@@ -83,6 +84,7 @@ let vm = new Vue({
             cluego_base_path: "",
             in: "",
             output: "",
+            output_name: "",
             type: null,
             species: null,
             limit: null,
@@ -122,9 +124,81 @@ let vm = new Vue({
         reanalysis_name: "",
         last_reanalysis_name: "",
         pine: null,
+        credits: [
+            {
+                "name": "Vue.js",
+                "url": "https://vuejs.org/",
+                "license_file": "vue.txt",
+                "license": null,
+            },
+            {
+                "name": "Font Awesome",
+                "url": "https://fontawesome.com/",
+                "license_file": "font-awesome.txt",
+                "license": null,
+            },
+            {
+                "name": "Electron",
+                "url": "https://electronjs.org/",
+                "license_file": "electron.txt",
+                "license": null,
+            },
+            {
+                "name": "electron-builder",
+                "url": "https://www.electron.build/",
+                "license_file": "electron-builder.txt",
+                "license": null,
+            },
+            {
+                "name": "Sass",
+                "url": "https://sass-lang.com/",
+                "license_file": "sass.txt",
+                "license": null,
+            },
+            {
+                "name": "PyInstaller",
+                "url": "https://www.pyinstaller.org/",
+                "license_file": "pyinstaller.txt",
+                "license": null,
+            },
+            {
+                "name": "Cytoscape",
+                "url": "https://cytoscape.org/",
+                "license_file": "cytoscape.txt",
+                "license": null,
+            },
+            {
+                "name": "py2cytoscape",
+                "url": "https://github.com/cytoscape/py2cytoscape",
+                "license_file": "py2cytoscape.txt",
+                "license": null,
+            },
+            {
+                "name": "GeneMANIA",
+                "url": "https://genemania.org/",
+            },
+            {
+                "name": "STRING",
+                "url": "https://string-db.org/",
+                "license_file": "string-db.txt",
+                "license": null,
+            },
+            {
+                "name": "ClueGO",
+                "url": "http://www.ici.upmc.fr/cluego/",
+                "license_file": "cluego.txt",
+                "license": null,
+            },
+            {
+                "name": "pandas",
+                "url": "https://pandas.pydata.org/",
+                "license_file": "pandas.txt",
+                "license": null,
+            },
+        ],
     },
     methods: {
-        run: async function(args) {
+        run: async function(args, is_reanalysis) {
             let that = this;
             if(!this.runnable() || this.running) {
                 return false;
@@ -133,7 +207,7 @@ let vm = new Vue({
 
             let file_check = this.runnable_file_check();
             if(!file_check["success"]) {
-                error_popup("File doesn't exist", `${file_check["file"]} doesn't exist.`);
+                error_popup("File error", `${file_check["file"]}`);
                 this.running = false;
                 return false;
             }
@@ -210,18 +284,23 @@ let vm = new Vue({
                     } else {
                         that.stdout += "Run failed\n";
                         that.stderr = stderr;
-                        /* delete the orphaned directory if it exists */
-                        if(new_session_dir && is_dir(new_session_dir)) {
-                            const dir_contents = fs.readdirSync(new_session_dir);
-                            if(dir_contents.length == 1 && dir_contents[0] === "PINE.log") {
-                                const pine_log_file = path.join(new_session_dir, "PINE.log");
-                                if(is_file(pine_log_file)) {
-                                    fs.unlinkSync(pine_log_file);
-                                    fs.rmdirSync(new_session_dir);
+                        http.get("http://localhost:1234/v1/commands/command/quit");
+
+                        if(!is_reanalysis) {
+                            /* delete the orphaned directory if it exists */
+                            if(new_session_dir && is_dir(new_session_dir)) {
+                                const dir_contents = fs.readdirSync(new_session_dir);
+                                if(dir_contents.length == 2 && dir_contents.includes("PINE.log") && dir_contents.includes("settings.json")) {
+                                    const pine_log_file = path.join(new_session_dir, "PINE.log");
+                                    const pine_settings_file = path.join(new_session_dir, "settings.json");
+                                    if(is_file(pine_log_file) && is_file(pine_settings_file)) {
+                                        fs.unlinkSync(pine_log_file);
+                                        fs.unlinkSync(pine_settings_file);
+                                        fs.rmdirSync(new_session_dir);
+                                    }
                                 }
                             }
                         }
-                        http.get("http://localhost:1234/v1/commands/command/quit");
                         resolve(false);
                         document.getElementById("log-wrapper").scrollTop = 0;
                     }
@@ -253,7 +332,7 @@ let vm = new Vue({
             }
 
             let args = this.pine_args();
-            this.run(args);
+            this.run(args, false);
         },
         run_with_cluego_subset: async function() {
             if(!this.runnable_with_cluego_subset()) {
@@ -267,12 +346,19 @@ let vm = new Vue({
             } else {
                 var reanalysis_name = this.generate_reanalysis_name();
             }
+
+            if(!is_dir(this.session_dir)) {
+                error_popup("Invalid session directory", "Your session directory may have been renamed or deleted.  Please reload the session or run this dataset again.", true);
+                return;
+            }
+
             if(!this.unique_reanalysis_name(reanalysis_name)) {
                 error_popup("Name in use", "This name has already been used for this session, please pick a different name", true);
                 return;
             }
             this.last_reanalysis_name = reanalysis_name;
             const filtered_file_name = path.join(this.session_dir, reanalysis_name + ".cluego.txt");
+            const log_file_name = path.join(this.session_dir, reanalysis_name + ".log");
             let file_data = [];
             file_data.push(this.cluego_pathways.header.join("\t"));
             for(const pathway of this.cluego_pathways.data) {
@@ -285,9 +371,12 @@ let vm = new Vue({
             let args = this.pine_args();
             args.push("-a");
             args.push(filtered_file_name);
-            let res = await this.run(args);
+            let res = await this.run(args, true);
             if(!res) {
                 fs.unlinkSync(filtered_file_name);
+                if(is_file(log_file_name)) {
+                    fs.unlinkSync(log_file_name);
+                }
             }
         },
         genemania_check: function() {
@@ -464,12 +553,23 @@ let vm = new Vue({
                 "--cluego-pval", this.input.cluego_pval,
                 "--reference-path", this.input.reference_path,
                 "--grouping", this.input.grouping,
-                "--cytoscape-executable", this.input.cytoscape_path,
-                "--enzyme", this.input.enzyme,
-                "--fasta-file", this.input.fasta_file,
-                "--mods", this.input.mods,
+                "--cytoscape-exe", this.input.cytoscape_path,
                 "--gui",
             ];
+            if(this.is_extra_options_required()) {
+                args.push("--mods");
+                args.push(this.input.mods);
+
+                args.push("--enzyme");
+                args.push(this.input.enzyme);
+
+                args.push("--fasta-file");
+                args.push(this.input.fasta_file);
+            }
+            if(this.input.output_name) {
+                args.push("--output-name");
+                args.push(this.input.output_name);
+            }
             if(this.input.significant) {
                 args.push("--significant");
             }
@@ -480,29 +580,42 @@ let vm = new Vue({
             return args;
         },
         runnable: function() {
-            if(this.input.in && this.get_cluego_mapping() && this.input.output && this.input.type && !this.running && this.validate_inputs()) {
+            const input_check =
+                this.input.in &&
+                this.get_cluego_mapping() &&
+                this.input.output &&
+                this.input.type &&
+                !this.running &&
+                this.validate_inputs();
+            if(input_check) {
                 if(this.is_extra_options_required()) {
-                    if(this.input.mods && this.input.fasta_file && this.input.enzyme) {
-                        return true;
-                    } else {
-                        return false;
+                    for(const erf of this.extra_required_fields) {
+                        if(!this.input[erf]) {
+                            return false;
+                        }
                     }
+                    return true;
                 } else {
                     return true;
                 }
             }
             return false;
         },
-        runnable_file_check: function() {
+        runnable_file_check: function(is_reanalysis) {
             let msg = null;
             if(!is_file(this.input.in)) {
-                msg = "Input file";
+                msg = "Input file doesn't exist.";
             } else if(!is_dir(this.input.output)) {
-                msg = "Output directory";
+                msg = "Output directory doesn't exist.";
             } else if(this.is_extra_options_required() && !is_file(this.input.fasta_file)) {
-                msg = "Fasta file";
+                msg = "Fasta file doesn't exist.";
             } else if(this.input.reference_path && !is_file(this.input.reference_path)) {
-                msg = "Reference path file";
+                msg = "Reference path file doesn't exist.";
+            } else if(this.input.output_name) {
+                const output_name_path = path.join(this.input.output, this.input.output_name);
+                if(is_file(output_name_path) || is_dir(output_name_path)) {
+                    msg = `Output path ${output_name_path} already exists.`;
+                }
             }
             if(msg) {
                 return {"success": false, "file": msg};
@@ -597,6 +710,7 @@ let vm = new Vue({
             this.input.mods = "S,T,Y";
             this.input.in = "";
             this.input.output = "";
+            this.input.output_name = "";
             this.input.remove_ambiguous = false;
         },
         read_cluego_pathways: function() {
@@ -604,7 +718,7 @@ let vm = new Vue({
 
             this.reset_cluego_pathways();
 
-            if(!this.session_exists() || !fs.existsSync(this.session_cluego_file) || !fs.statSync(this.session_cluego_file).isFile()) {
+            if(!this.session_exists() || !is_file(this.session_cluego_file)) {
                 return;
             }
 
@@ -671,24 +785,37 @@ let vm = new Vue({
             }
         },
         save_settings: function(settings_file) {
-            fs.writeFileSync(settings_file, JSON.stringify(this.input, null, 4));
+            let settings = {};
+            for(const inp in this.input) { 
+                if(inp === "output_name") {
+                    continue;
+                }
+                if(!this.is_extra_options_required() && this.extra_required_fields.includes(inp)) {
+                    continue;
+                }
+                settings[inp] = this.input[inp];
+            }
+            fs.writeFileSync(settings_file, JSON.stringify(settings, null, 4));
         },
         load_settings: function(settings_file) {
             if(!is_file(settings_file)) {
                 return;
             }
             const raw = fs.readFileSync(settings_file, "utf8");
-            const saved_input = JSON.parse(raw);
-            for(const key in saved_input) {
-                if(key in this.input) {
-                    if(key === "cluego_base_path") {
-                        this.setCluegoBasePath(saved_input[key], false);
-                    } else if(key === "cytoscape_path") {
-                        this.setCytoscapePath(saved_input[key], false);
-                    } else {
-                        Vue.set(this.input, key, saved_input[key]);
+            try {
+                const saved_input = JSON.parse(raw);
+                for(const key in saved_input) {
+                    if(key in this.input) {
+                        if(key === "cluego_base_path") {
+                            this.setCluegoBasePath(saved_input[key], false);
+                        } else if(key === "cytoscape_path") {
+                            this.setCytoscapePath(saved_input[key], false);
+                        } else {
+                            Vue.set(this.input, key, saved_input[key]);
+                        }
                     }
                 }
+            } catch(e) {
             }
         },
         get_settings_file: function() {
@@ -1103,7 +1230,7 @@ let vm = new Vue({
             if(!this.session_dir) {
                 return "";
             }
-            return path.join(this.session_dir, "settings.txt");
+            return path.join(this.session_dir, "settings-gui.json");
         },
         selectable_species: function() {
             let selectable = [];
@@ -1138,6 +1265,10 @@ let vm = new Vue({
         },
         "cluego_pathways.ontology_sources_filter": function() {
             this.cluego_pathways.page = 1;
+        },
+        "input.output_name": function(new_val) {
+            let re = /[^a-zA-Z0-9-_]/g;
+            this.$set(this.input, "output_name", new_val.replace(re, ""));
         },
     }
 });
